Finish up core functionality
This fleshes out the unpacking of MSI files a bit to account for
directories, as well as implementing packing into an output directory
that can then be used either as is, or tarballed etc.

Initially was going to add support for generating tarballs directly,
but that's really not needed since the user can just do that themselves.